requirements.txt
openai==0.27.2
pinecone-client==2.2.1
python-dotenv==1.0.0
pgvector==0.1.6
SQLAlchemy==2.0.9
psycopg2==2.9.6
faiss-cpu==1.7.3